+++ changelog.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [unreleased]
 
+- Remove gregoryv/workdir dependency
 - Update gregoryv/english dependency
 
 ## [0.22.0] 2021-10-23